Beginning refactor into separate files
